Update serialconnectioncheck.py
This file Initializes the 'ewart candapter' and then opens the port for communication, it recieves all the data that the candapter sends. We read approximately 189 bytes and each data variable is seperated by a "t".So splitting the string array accordingly, we get the required parameters to decode the received string and obtain the mppt parameter.